requirements.txt
opensmile
pandas
langchain
openai
python-dotenv